Refactor to extract Parser class and move Storage to its own package.
