+++ src/main/java/odin/storage/Storage.java
@@ -1,5 +1,6 @@
-package odin;
+package odin.storage;
 
+import odin.IllegalFileException;
 import odin.task.Deadline;
 import odin.task.Event;
 import odin.task.Task;
